PCHR-3941: Use RequireJS in HR UI
