+++ hrui/js/build.js
@@ -1,0 +1,9 @@
+/* eslint-disable */
+
+({
+  baseUrl: 'src',
+  out: 'dist/hrui.min.js',
+  name: 'hrui',
+  skipModuleInsertion: true,
+  generateSourceMaps: true
+})
